Fixed warnings to do with the previous commit's transition to making m_population an unsigned long long int

+++ GeoRegions/Region.cpp
@@ -138,9 +138,9 @@ std::string Region::getRegionLabel() const {
     return regionLabel(getType());
 }
 
-unsigned int Region::computeTotalPopulation() {
+unsigned long long int Region::computeTotalPopulation() {
     // TODO: test me
-    unsigned int population = m_population;
+    unsigned long long int population = m_population;
     for (auto &&r : m_subRegions) {
         if (r != nullptr) {
             population += r->getPopulation();
@@ -163,7 +163,7 @@ void Region::displayAll(std::ostream &out, unsigned int displayLevel) {
         out << std::setw(displayLevel * TAB_SIZE) << " ";
     }
 
-    unsigned totalPopulation = computeTotalPopulation();
+    unsigned long long totalPopulation = computeTotalPopulation();
     double area = getArea();
     double density = (double) totalPopulation / area;
 
@@ -184,7 +184,7 @@ void Region::displaySubLevel(std::ostream &out, unsigned int displayLevel, bool 
         out << std::setw(displayLevel * TAB_SIZE) << " ";
     }
 
-    unsigned totalPopulation = computeTotalPopulation();
+    unsigned long long totalPopulation = computeTotalPopulation();
     double area = getArea();
     double density = (double) totalPopulation / area;
